% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSK.R
\name{design_MVSK_portfolio}
\alias{design_MVSK_portfolio}
\title{Design high-order portfolio based on weighted linear combination of first four moments}
\usage{
design_MVSK_portfolio(
  lmd = rep(1, 4),
  X_moments,
  w_init = rep(1/length(X_moments$mu), length(X_moments$mu)),
  leverage = 1,
  method = c("Q-MVSK", "MM", "DC"),
  tau_w = 0,
  gamma = 1,
  zeta = 1e-08,
  maxiter = 100,
  ftol = 1e-05,
  wtol = 1e-04,
  stopval = -Inf
)
}
\arguments{
\item{lmd}{a numerical vector of length 4, indicating the weights of first four moments.}

\item{X_moments}{a list of moment parameters, see \code{\link{estimate_moments}}.}

\item{w_init}{a numerical vector, indicating the initial value of portfolio weights.}

\item{leverage}{a number (>= 1), indicating the leverage of portfolio.}

\item{method}{a string, indicating the algorithm method, must be one of: "Q-MVSK", "MM", "DC".}

\item{tau_w}{a number (>= 0), guaranteeing the strong convexity of approximating function.}

\item{gamma}{a number (0 < gamma <= 1), indicating the initial value of gamma.}

\item{zeta}{a number (0 < zeta < 1), indicating the diminishing paramater of gamma.}

\item{maxiter}{a positive integer, setting the maximum iteration.}

\item{ftol}{a positive number, setting the convergence criterion of function objective.}

\item{wtol}{a positive number, setting the convergence criterion of portfolio weights.}

\item{stopval}{a number, setting the stop value of objective.}
}
\value{
A list containing the following elements:
\item{\code{w}}{Optimal portfolio vector.}
\item{\code{cpu_time}}{Time usage with iteration.}
\item{\code{objs}}{Function objective with iteration.}
\item{\code{convergence}}{Bloolean flag to indicate whether or not the optimization converged.}
\item{\code{moments}}{Moments of portfolio return at optimal portfolio weights.}
}
\description{
Design high-order portfolio based on weighted linear combination of first four moments
(i.e., mean, variance, skewness, and kurtosis):
\preformatted{
  minimize     - lmd1*(w'*mu) + lmd2*(w'*Sigma*w) 
               - lmd3*(w'*Phi*w*w) + lmd4*(w'*Psi*w*w*w)
  subject to   ||w||_1 <= leverage, sum(w) == 1,
}
}
\examples{
\dontrun{
library(highOrderPortfolios)
data(X50)

# estimate moments
X_moments <- estimate_moments(X50)

# decide moment weights
xi <- 10
lmd <- c(1, xi/2, xi*(xi+1)/6, xi*(xi+1)*(xi+2)/24)

# portfolio optimization
sol <- design_MVSK_portfolio(lmd, X_moments)
}

}
\references{
Rui Zhou and and Daniel P. Palomar, “Solving High-Order Portfolios via Successive Convex Approximation Algorithms,”
under review, 2020.
}
\author{
Rui Zhou and Daniel P. Palomar
}
