% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{estimate_moments}
\alias{estimate_moments}
\title{Estimate first four moment parameters of multivariate observations}
\usage{
estimate_moments(X, adjust_magnitude = FALSE)
}
\arguments{
\item{X}{a data matrix}

\item{adjust_magnitude}{a boolean variable, indicating whether to adjust the order of magnitude of parameters.
NOTE: It is specially designed for \code{\link{MVSKtilting}} function.}
}
\value{
A list containing the following elements:
\item{\code{mu}}{mean vector.}
\item{\code{Sgm}}{Covariance matrix.}
\item{\code{Phi}}{Coskewness matrix in vector form (collecting only the unique elements).}
\item{\code{Psi}}{Cokurtosis matrix in vector form (collecting only the unique elements).}
\item{\code{Phi_shred}}{Partition on \code{Phi}, see reference.}
\item{\code{Psi_shred}}{Partition on \code{Psi}, see reference.}
}
\description{
Estimate first four moments of multivariate observations, namely,
mean vector, covariance matrix, coskewness matrix, and cokurtosis matrix.
}
\examples{
\dontrun{
library(highOrderPortfolios)
data(X50)

X_moments <- estimate_moments(X50)
}

}
\references{
Rui Zhou and and Daniel P. Palomar, “Solving High-Order Portfolios via Successive Convex Approximation Algorithms,”
under review, 2020.
}
\author{
Rui Zhou and Daniel P. Palomar
}
