% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_MVSK_portfolio.R
\name{estimate_moments_MVSK}
\alias{estimate_moments_MVSK}
\title{Estimate first four moments of multivariate observations}
\usage{
estimate_moments_MVSK(X)
}
\description{
Estimate first four moments of multivariate observations, namely,
mean, variance, skewness, and kurtosis.
}
\examples{
library(MVSKportfolio)
data(X50)

estimate_moments_MVSK(X50)

}
\references{
Rui Zhou and and Daniel P. Palomar, “Solving High-Order Portfolios via Successive Convex Approximation Algorithms,”
under review, 2020.
}
\author{
Rui Zhou and Daniel P. Palomar
}
