% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{evalMoms}
\alias{evalMoms}
\title{Evaluate first four moments of a given portfolio}
\usage{
evalMoms(w, mom_params)
}
\arguments{
\item{w}{a numerical vector as portfolio weights.}

\item{mom_params}{a list of moment parameters, see \code{\link{estMomParams}}.}
}
\value{
Four moments of the given portfolio.
}
\description{
Estimate first four moments of a given portfolio's return, namely,
mean, variance, skewness, and kurtosis.
}
\examples{
\dontrun{
library(highOrderPortfolios)
data(X50)

params <- estimate_moments_MVSK(X50)
moments <- evalMoms(rep(1/50, 50), params)
}

}
\references{
Rui Zhou and and Daniel P. Palomar, “Solving High-Order Portfolios via Successive Convex Approximation Algorithms,”
under review, 2020.
}
\author{
Rui Zhou and Daniel P. Palomar
}
