% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSKtilting.R
\name{MVSKtilting}
\alias{MVSKtilting}
\title{Design high-order portfolio by tilting a given portfolio to the MVSK efficient frontier}
\usage{
MVSKtilting(
  d = rep(1, 4),
  mom_params,
  w_init = rep(1/length(mom_params$mu), length(mom_params$mu)),
  w0 = w_init,
  moms0 = NULL,
  leverage = 1,
  kappa = 0,
  method = c("Q-MVSKT", "L-MVSKT"),
  tau_w = 1e-05,
  tau_delta = 1e-05,
  gamma = 1,
  zeta = 1e-08,
  maxiter = 100,
  ftol = 1e-05,
  wtol = 1e-05,
  theta = 0.5,
  stopval = -Inf
)
}
\arguments{
\item{d}{a numerical vector of length 4, indicating the tilting direction of first four moments.}

\item{mom_params}{a list of moment parameters, see \code{\link{estMomParams}}.}

\item{w_init}{a numerical vector, indicating the initial value of portfolio weights.}

\item{w0}{a numerical vector, indicating the reference portfolio vector.}

\item{moms0}{a numerical vector, indicating the reference moments.}

\item{leverage}{a number (>= 1), indicating the leverage of portfolio.}

\item{kappa}{a number, indicating the maximum tracking error volatility.}

\item{method}{a string, indicating the algorithm method, must be one of: "Q-MVSK", "MM", "DC".}

\item{tau_w}{a number (>= 0), guaranteeing the strong convexity of approximating function.}

\item{tau_delta}{a number (>= 0), guaranteeing the strong convexity of approximating function.}

\item{gamma}{a number (0 < gamma <= 1), indicating the initial value of gamma.}

\item{zeta}{a number (0 < zeta < 1), indicating the diminishing paramater of gamma.}

\item{maxiter}{a positive integer, setting the maximum iteration.}

\item{ftol}{a positive number, setting the convergence criterion of function objective.}

\item{wtol}{a positive number, setting the convergence criterion of portfolio weights.}

\item{theta}{a number (0 < theta < 1), setting the combination coefficient when enlarge feasible set.}

\item{stopval}{a number, setting the stop value of objective.}
}
\value{
A list containing the following elements:
\item{\code{w}}{Optimal portfolio vector.}
\item{\code{times}}{Time usage with iteration.}
\item{\code{objs}}{Function objective with iteration.}
\item{\code{convergence}}{Bloolean flag to indicate whether or not the optimization converged.}
\item{\code{moments}}{Moments of portfolio return at optimal portfolio weights.}
}
\description{
Design high-order portfolio by tilting a given portfolio to the MVSK efficient frontier
(i.e., mean, variance, skewness, and kurtosis):
\preformatted{
  minimize     delta
               m1(w) >= m1(w0) + delta*d1
               m2(w) <= m2(w0) - delta*d2
               m3(w) >= m3(w0) + delta*d3
               m4(w) <= m4(w0) - delta*d4
               (w-w0)'Sigma(w-w0) <= kappa^2
  subject to   ||w||_1 <= leverage, sum(w) == 1,
}
}
\examples{
\dontrun{
library(highOrderPortfolios)
data(X50)

# estimate parameters
params <- estMomParams(X50, align_order = TRUE)

# decide problem setting
w0 <- rep(1/50, 50)
moms0 <- evalMoms(w = w0, mom_params = params)
d <- abs(moms0) 
kappa <- sqrt(w0\%*\%params$Sgm\%*\%w0) * 0.3

# portfolio optimization
sol <- MVSKtilting(d = d, mom_params = params, w_init = w0, w0 = w0, moms0 = moms0, kappa = kappa)
}

}
\references{
Rui Zhou and and Daniel P. Palomar, “Solving High-Order Portfolios via Successive Convex Approximation Algorithms,”
under review, 2020.
}
\author{
Rui Zhou and Daniel P. Palomar
}
