% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X100.R
\docType{data}
\name{X100}
\alias{X100}
\title{Synthetic 500x100 matrix dataset}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 500 rows and 100 columns.
}
\usage{
data(X100)
}
\description{
Synthetic 500x100 matrix dataset containing 500 realizations
of 100 variables.
}
\keyword{dataset}
