% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X200.R
\docType{data}
\name{X200}
\alias{X200}
\title{Synthetic 1000x200 matrix dataset}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 1000 rows and 200 columns.
}
\usage{
data(X200)
}
\description{
Synthetic 1000x200 matrix dataset containing 1000 realizations
of 200 variables.
}
\keyword{dataset}
