% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X50.R
\docType{data}
\name{X50}
\alias{X50}
\title{Synthetic 300x50 matrix dataset}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 503 rows and 50 columns.}
\usage{
data(X50)
}
\description{
Synthetic 300x50 matrix dataset containing 300 realizations
of 50 variables.
}
\keyword{dataset}
