% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_MVSK_portfolio.R
\name{design_MVSK_portfolio}
\alias{design_MVSK_portfolio}
\title{Design high-order portfolio based on weighted linear combination of first four moments}
\usage{
design_MVSK_portfolio()
}
\description{
Design high-order portfolio based on weighted linear combination of first four moments
(i.e., mean, variance, skewness, and kurtosis):
\preformatted{
  maximize     w'*mu
  subject to   w'*Sigma*w <= var_max, 
               ||w||_1 <= leverage,  (w >= 0),
               (beta'*w = 0),
}
where \code{beta} can be a vector or a matrix.
}
\examples{
library(MVSKportfolio)
data(X50)

moments <- estimate_moments_MVSK(X50)
w <- design_MVSK_portfolio(moments)

}
\references{
Rui Zhou and and Daniel P. Palomar, “Solving High-Order Portfolios via Successive Convex Approximation Algorithms,”
under review, 2020.
}
\author{
Rui Zhou and Daniel P. Palomar
}
