% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{estimate_moments}
\alias{estimate_moments}
\title{Estimate first four moment parameters of multivariate observations}
\usage{
estimate_moments(X, adjust_magnitude = FALSE)
}
\arguments{
\item{X}{Data matrix.}

\item{adjust_magnitude}{Boolean indicating whether to adjust the order of magnitude of parameters.
NOTE: It is specially designed for \code{\link{design_MVSKtilting_portfolio}} function.}
}
\value{
A list containing the following elements:
\item{\code{mu}}{Mean vector.}
\item{\code{Sgm}}{Covariance matrix.}
\item{\code{Phi}}{Co-skewness matrix in vector form (collecting only the unique elements).}
\item{\code{Psi}}{Co-kurtosis matrix in vector form (collecting only the unique elements).}
\item{\code{Phi_shred}}{Partition on \code{Phi} (see reference).}
\item{\code{Psi_shred}}{Partition on \code{Psi} (see reference).}
}
\description{
Estimate first four moments of multivariate observations, namely,
mean vector, covariance matrix, coskewness matrix, and cokurtosis matrix.
}
\examples{
\donttest{
library(highOrderPortfolios)
data(X50)

X_moments <- estimate_moments(X50)
}

}
\references{
Rui Zhou and and Daniel P. Palomar, “Solving High-Order Portfolios via Successive Convex Approximation Algorithms,”
under review, 2020. <https://arxiv.org/abs/2008.00863>
}
\author{
Rui Zhou and Daniel P. Palomar
}
