% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{eval_portfolio_moments}
\alias{eval_portfolio_moments}
\title{Evaluate first four moments of a given portfolio}
\usage{
eval_portfolio_moments(w, X_moments)
}
\arguments{
\item{w}{Numerical vector with portfolio weights.}

\item{X_moments}{List of moment parameters, as obtained with function \code{\link{estimate_moments}}.}
}
\value{
Four moments of the given portfolio.
}
\description{
Estimate first four moments of a given portfolio's return, namely,
mean, variance, skewness, and kurtosis.
}
\examples{
\donttest{
library(highOrderPortfolios)
data(X50)

X_moments <- estimate_moments(X50)
w_moments <- eval_portfolio_moments(rep(1/50, 50), X_moments)
}

}
\references{
Rui Zhou and and Daniel P. Palomar, “Solving High-Order Portfolios via Successive Convex Approximation Algorithms,”
under review, 2020. <https://arxiv.org/abs/2008.00863>
}
\author{
Rui Zhou and Daniel P. Palomar
}
