% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{eval_portfolio_moments}
\alias{eval_portfolio_moments}
\title{Evaluate first four moments of a given portfolio}
\usage{
eval_portfolio_moments(w, X_moments)
}
\arguments{
\item{w}{Numerical vector with portfolio weights.}

\item{X_moments}{List of moment parameters, as obtained with function \code{\link{estimate_moments}}.}
}
\value{
Four moments of the given portfolio.
}
\description{
Estimate first four moments of a given portfolio's return, namely,
mean, variance, skewness, and kurtosis.
}
\examples{

library(highOrderPortfolios)
data(X50)

X_moments <- estimate_moments(X50[, 1:10])
w_moments <- eval_portfolio_moments(rep(1/10, 10), X_moments)


}
\references{
R. Zhou and D. P. Palomar, "Solving High-Order Portfolios via Successive Convex Approximation Algorithms," 
in \emph{IEEE Transactions on Signal Processing}, vol. 69, pp. 892-904, 2021.
<doi:10.1109/TSP.2021.3051369>.
}
\author{
Rui Zhou and Daniel P. Palomar
}
