% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highOrderPortfolios-package.R
\docType{package}
\name{highOrderPortfolios-package}
\alias{highOrderPortfolios-package}
\title{highOrderPortfolios: Design of High-Order Portfolios via Mean, Variance, Skewness, and Kurtosis}
\description{
The classical Markowitz's mean-variance portfolio formulation ignores 
heavy tails and skewness. High-order portfolios use higher order moments to
better characterize the return distribution. Different formulations and fast 
algorithms are proposed for high-order portfolios based on the mean, variance, 
skewness, and kurtosis.
}
\section{Functions}{

\code{\link{design_MVSK_portfolio}} and \code{\link{design_MVSKtilting_portfolio}}
}

\section{Help}{

For a quick help see the README file:
\href{https://github.com/dppalomar/highOrderPortfolios/blob/master/README.md}{GitHub-README}.
}

\references{
R. Zhou and D. P. Palomar, "Solving High-Order Portfolios via Successive Convex Approximation Algorithms," 
in \emph{IEEE Transactions on Signal Processing}, vol. 69, pp. 892-904, 2021.
<doi:10.1109/TSP.2021.3051369>.
}
\author{
Rui Zhou and Daniel P. Palomar
}
