% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X50.R
\docType{data}
\name{X50}
\alias{X50}
\title{Synthetic 250x50 matrix dataset}
\format{
An object of class \code{matrix} with 250 rows and 50 columns.
}
\usage{
data(X50)
}
\description{
Synthetic 250x50 matrix dataset containing 250 realizations
of 50 variables.
}
\keyword{dataset}
